+++ include/vrv/tie.h
@@ -36,6 +36,7 @@ public:
      */
     ///@{
     Tie();
+    Tie(const std::string &classid);
     virtual ~Tie();
     virtual Object *Clone() const { return new Tie(*this); }
     virtual void Reset();
@@ -51,7 +52,7 @@ public:
     virtual TimeSpanningInterface *GetTimeSpanningInterface() { return dynamic_cast<TimeSpanningInterface *>(this); }
     ///@}
 
-    bool CalculatePosition(Doc *doc, Staff *staff, int x1, int x2, int spanningType, Point bezier[4]);
+    virtual bool CalculatePosition(Doc *doc, Staff *staff, int x1, int x2, int spanningType, Point bezier[4]);
 
     //----------//
     // Functors //