+++ include/vrv/view.h
@@ -206,7 +206,8 @@ protected:
     void DrawBarLineDots(DeviceContext *dc, Staff *staff, BarLine *barLine);
     void DrawLedgerLines(DeviceContext *dc, Staff *staff, ArrayOfLedgerLines *lines, bool below, bool cueSize);
     void DrawMeasure(DeviceContext *dc, Measure *measure, System *system);
-    void DrawMNum(DeviceContext *dc, MNum *mnum, Measure *measure);
+    void DrawMeterSigGrp(DeviceContext *dc, Layer *layer, Staff *staff);
+    void DrawMNum(DeviceContext *dc, MNum *mnum, Measure *measure, int yOffset);
     void DrawStaff(DeviceContext *dc, Staff *staff, Measure *measure, System *system);
     void DrawStaffLines(DeviceContext *dc, Staff *staff, Measure *measure, System *system);
     void DrawLayer(DeviceContext *dc, Layer *layer, Staff *staff, Measure *measure);