+++ src/ligature.cpp
@@ -69,32 +69,41 @@ bool Ligature::IsSupportedChild(Object *child)
     return true;
 }
 
+Note *Ligature::GetFirstNote()
+{
+    const ArrayOfObjects *childList = this->GetList(this); // make sure it's initialized
+    assert(childList->size() > 0);
+
+    Note *firstNote = vrv_cast<Note *>(childList->front());
+    assert(firstNote);
+    return firstNote;
+}
+
+Note *Ligature::GetLastNote()
+{
+    const ArrayOfObjects *childList = this->GetList(this); // make sure it's initialized
+    assert(childList->size() > 0);
+
+    // The first note is the bottom
+    Note *lastNote = vrv_cast<Note *>(childList->back());
+    assert(lastNote);
+    return lastNote;
+}
+
 void Ligature::FilterList(ArrayOfObjects *childList)
 {
     // Retain only note children of ligatures
     ArrayOfObjects::iterator iter = childList->begin();
 
     while (iter != childList->end()) {
-        if (!(*iter)->IsLayerElement()) {
+        if (!(*iter)->Is(NOTE)) {
             // remove anything that is not an LayerElement
-            iter = childList->erase(iter);
-            continue;
-        }
-        LayerElement *currentElement = vrv_cast<LayerElement *>(*iter);
-        assert(currentElement);
-        if (!currentElement->HasInterface(INTERFACE_DURATION)) {
             iter = childList->erase(iter);
         }
         else {
-            Note *n = dynamic_cast<Note *>(currentElement);
-
-            if (n) {
-                ++iter;
-            }
-            else {
-                // if it is not a note, drop it
-                iter = childList->erase(iter);
-            }
+            // assert that we keep only notes
+            assert(dynamic_cast<Note *>(*iter));
+            ++iter;
         }
     }
 
@@ -121,6 +130,8 @@ int Ligature::CalcLigatureNotePos(FunctorParams *functorParams)
 {
     FunctorDocParams *params = vrv_params_cast<FunctorDocParams *>(functorParams);
     assert(params);
+
+    if (params->m_doc->GetOptions()->m_ligatureAsBracket.GetValue()) return FUNCTOR_CONTINUE;
 
     m_drawingShapes.clear();